// src/api/jr/types.d.ts
import { TimeStamps } from "@/shared/types"

export interface DepartmentRequestType extends TimeStamps {
	departmentRequestID: string
	departmentRequestName: string
	address: string
}

export type DepartmentRequestPayload = {
	departmentRequestName: string
	address: string
}

export interface PostResponse {
	data: string
	status: number
	message: string
}

export interface UpdateDepartmentRequestType {
	departmentRequestID: string
	departmentRequestName: string
	address: string
}

export interface DivisionManagerType extends TimeStamps {
	divisionManagerID: string
	firstName: string
	lastName: string
	phoneNumber: string
	email: string
	position: string
	lineCode: string
	lineUserID: string
}

export interface UpdateDivisionManagerType {
	divisionManagerID: string
	firstName: string
	lastName: string
	phoneNumber: string
	email: string
	position: string
	lineCode: string
	lineUserID: string
}

export type DivisionManagerPayload = {
	firstName: string
	lastName: string
	phoneNumber: string
	email: string
	position: string
	lineCode: string
	lineUserID?: string
}

export interface AddDepartmentResponse {
	data: string
	status: number
	message: string
}


export interface JobPositionType extends TimeStamps {
	jobPositionID: string
	divisionManagerID: string
	jobDescription: string
	positionName: string
	jobLevel: string
	basicQuestionList: BasicQuestionType[]
	masterQuestionsList: MasterQuestionType[]
}

export type MasterQuestion = {
	questionThai: string
	questionEnglish: string
}

export type BasicQuestion = {
	question: string
	questionTh: string
}

export interface AddJobPositionPayloadType {
	divisionManagerID: string
	jobDescription: string
	positionName: string
	jobLevel: string
	addQuestionList?: MasterQuestion[] | undefined
	addBasicQuestionList?: BasicQuestion[] | undefined
}

export interface EditMasterQuestionPayloadType {
	questionID:string
	jobPositionID: string
	updateQuestionList?: MasterQuestion[] | undefined
	updateBasicQuestionList?: BasicQuestionType[] | undefined
}

export type BasicQuestionType = {
	basicQuestionID?: string
	questionID?:string
	jobPositionID?: string
	question: string
	questionTh: string
	createdAt?: string
	updatedAt?: string
}

export interface UpdateBasicQuestionPayloadType {
	basicQuestionID?:string
	jobPositionID: string
	questionTh:string
	question:string
}

export type MasterQuestionType = {
	questionID: number
	jobPositionID: string
	questionThai: string
	questionEnglish: string
	createdAt: string
	updatedAt: string
}


export interface UpdateMasterQuestionPayloadType {
	questionID:number
	jobPositionID: string
	questionThai:string
	questionEnglish:string
}